Added typecasts to eliminate gcc printf format warnings on OpenBSD (time_t is not unsigned long).

+++ sbbs3/logon.cpp
@@ -2,7 +2,7 @@
 
 /* Synchronet user logon routines */
 
-/* $Id: logon.cpp,v 1.29 2002/11/13 03:08:00 rswindell Exp $ */
+/* $Id: logon.cpp,v 1.30 2003/01/03 20:52:48 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
@@ -115,7 +115,7 @@ bool sbbs_t::logon()
 			,cfg.cdt_per_dollar*cfg.node_dollars_per_call);
 		sprintf(str,"%s #%u was billed $%d T: %lu seconds"
 			,useron.alias,useron.number
-			,cfg.node_dollars_per_call,now-answertime);
+			,cfg.node_dollars_per_call,(ulong)(now-answertime));
 		logline("$+",str);
 		hangup();
 		return(false);
